refactor: 모달 공통 컴포넌트를 분리
